ws_client/requirements.txt
websocket-client==1.5.1
getmac==0.9.3
